BOX-97: refactor(card-edit): refine card-pages structure
