+++ src/pages/card/README.md
@@ -1,0 +1,4 @@
+# Страницы для работы с карточками
+
+- `view` - Страница просмотра карточки (основная)
+- `edit` - Страница редактирования карточки
