+++ src/pages/card/index.ts
@@ -1,0 +1,2 @@
+export { CardEditPage } from './edit';
+export { CardPage } from './view';
