+++ src/pages/card/view/model.ts
